third_party/WebKit/LayoutTests/http/tests/devtools/console/console-link-to-snippet-expected.txt

CONSOLE MESSAGE: line 1: 239
CONSOLE ERROR: line 2: 42
Test that link to snippet works.


Running: testConsoleLogAndReturnMessageLocation
name1:1 239
name1:1 42

Running: testSnippetSyntaxError
Line Message was added: name2 Error 'Uncaught SyntaxError: Unexpected token }':1:1
name2:2 Uncaught SyntaxError: Unexpected token }

Running: testConsoleErrorHighlight
Line Message was added: name3 Error '42':1:10
name3:2 42
(anonymous) @ name3:2
